pyproject.toml
[tool.poetry]
name = "matching"
version = "0.1.0"
description = ""
authors = ["anon"]

[tool.poetry.dependencies]
python = "^3.7.1"
numpy = "1.20.0"
pandas = "^1.2.3"
sklearn = "^0.0"
scikit-learn = "^0.24.1"
hydra-core = "1.1.0.dev4"
wandb = "^0.10.23"
matplotlib = "^3.4.1"
seaborn = "^0.11.1"
plotly = "^4.14.3"
torch = "^1.8.1"
tqdm = "^4.60.0"
pretrainedmodels = "0.7.4"
gdown = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
